refactor(features): reorganize feature-related classes into features package
Move Feature, FeatureMetadata and FeatureManager to features package for better organization. Update all imports accordingly. Also improve Feature interface by changing setting() return type to Optional<Dialog> for better null safety.
