+++ src/mindustrytool/features/FeatureMetadata.java
@@ -1,4 +1,4 @@
-package mindustrytool;
+package mindustrytool.features;
 
 public record FeatureMetadata(String name, String description, String icon) {
     public FeatureMetadata(String name, String description) {
