+++ src/mindustrytool/features/browser/map/MapBrowserFeature.java
@@ -5,9 +5,9 @@ import arc.Events;
 import mindustry.Vars;
 import mindustry.game.EventType.ClientLoadEvent;
 import mindustry.gen.Icon;
-import mindustrytool.Feature;
-import mindustrytool.FeatureMetadata;
-import mindustrytool.FeatureManager;
+import mindustrytool.features.Feature;
+import mindustrytool.features.FeatureManager;
+import mindustrytool.features.FeatureMetadata;
 
 public class MapBrowserFeature implements Feature {
     private MapDialog mapDialog;
